reorganized mapper and deleted bloat
