+++ cusub_cortex/mapper/scripts/sim_truth_pub.py
@@ -19,7 +19,7 @@ class SimTruthPub:
     def __init__(self):
         rospy.loginfo("Sim Truth Publisher Initializing")
         objects = rospy.get_param('sim_pub/objects_to_publish')
-        self.pub = rospy.Publisher('cusub_cortex/mapper_in/task_poses', Detection, queue_size=1)
+        self.pub = rospy.Publisher('cusub_perception/mapper/task_poses', Detection, queue_size=1)
 
         self.objs = {}        
         for obj_name in objects:
