--- cusub_cortex/mapper/setup.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-from distutils.core import setup
-from catkin_pkg.python_setup import generate_distutils_setup
-
-d = generate_distutils_setup()
-d['packages'] = ['pose_filter']
-d['package_dir'] = {'': 'src'}
-
-setup(**d)
