+++ cusub_cortex/state_machine/config/mapper.yaml
@@ -1,0 +1,1 @@
+../../mapper/config/mapper.yaml
